bicycle/requirements.txt
asgiref==3.4.1
async-timeout==4.0.3
Django==4.0
django-autoreload==0.1.1
django-cors-headers==4.3.0
django-redis==5.4.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
mysqlclient==2.2.0
Pillow==10.1.0
pyinotify==0.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
websocket-server==0.6.4
